Dev #600 - search data elements
