+++ app/controllers/search_controller.rb
@@ -16,11 +16,9 @@ class SearchController < ApplicationController
 private
 
   def search
-    no_category_id = Category.find_by(name: 'No Category')&.id
     PgSearch.multisearch(search_params[:search])
-            .where(searchable_type: 'Concept')
-            .where.not(searchable_name: 'No Concept', searchable_category_id: no_category_id)
-            .includes(searchable: %i[data_elements category])
+            .where(searchable_type: 'DataElement')
+            .includes(searchable: %i[concept])
   end
 
   def sorted_search
@@ -53,24 +51,24 @@ private
   end
 
   def filters
-    @filters ||= params.permit(filter: { category_id: [], tab_name: [], years: [] })&.dig(:filter) || {}
+    @filters ||= params.permit(filter: { concept_id: [], tab_name: [], years: [] })&.dig(:filter) || {}
   end
 
   def filter_results(results)
-    results = filter_categories(results, filters.dig(:category_id))
+    results = filter_concepts(results, filters.dig(:concept_id))
     results = filter_years(results, filters.dig(:years))
     filter_tab_names(results, filters.dig(:tab_name))
   end
 
   def build_filters
-    @categories, data_elements, active_data_elements = build_categories_and_data_elements
-    @years, @tabs = build_years_tabs(data_elements, active_data_elements)
+    @concepts = build_concepts
+    @years, @tabs = build_years_tabs
   end
 
-  def filter_categories(results, category_ids)
-    return results if category_ids.blank?
+  def filter_concepts(results, concept_ids)
+    return results if concept_ids.blank?
 
-    results.where(searchable_category_id: category_ids)
+    results.where(searchable_category_id: concept_ids)
   end
 
   def filter_years(results, years)
@@ -93,27 +91,22 @@ private
     results.where.overlap(searchable_tab_names: tab_names)
   end
 
-  def build_categories_and_data_elements
-    categories = []
-    data_elements = []
-    active_data_elements = []
+  def build_concepts
+    concepts = []
     active_search = filtered_search.map(&:searchable).map(&:id).uniq.sort
     sorted_search.map(&:searchable).each do |searchable|
       is_active = active_search.include?(searchable.id)
-      categories.push(category: searchable.category, active: is_active)
-      data_elements.push(searchable.data_elements)
-      active_data_elements.push(searchable.data_elements) if is_active
+      concepts.push(concept: searchable.concept, active: is_active)
     end
-    [categories.uniq { |c| c[:category] }.sort { |a, b| a[:category].name <=> b[:category].name },
-     data_elements.flatten.uniq, active_data_elements.flatten.uniq]
+    concepts.uniq { |c| c[:concept] }.sort { |a, b| a[:concept].name <=> b[:concept].name }
   end
 
-  def build_years_tabs(data_elements, active_data_elements)
+  def build_years_tabs
     years = []
     tabs = []
-    active_tabs = active_data_elements.map(&:datasets).map(&:to_a).flatten.map(&:tab_name).compact.uniq
-    data_elements.each do |de|
-      years.push(collect_years(de, active_data_elements.include?(de)))
+    active_tabs = @filtered_search.map(&:searchable).map(&:datasets).map(&:to_a).flatten.map(&:tab_name).compact.uniq
+    @sorted_search.map(&:searchable).each do |de|
+      years.push(collect_years(de, @filtered_search.include?(de)))
       de.datasets.map(&:tab_name).each do |tab|
         tabs.push(tab: tab, active: active_tabs.include?(tab))
       end
