+++ app/models/concerns/indexing.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+# Indexing module for PostgreSQL and PGSearch
+module Indexing; end
